Day 7 Task/requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain_google_genai
pypdf2
faiss-cpu
pandas
openpyxl